Completed support for MISP_EMUL, added multi-threaded fast-forwarding for it. It now correctly executes all SPEC CPU 2000 and Splash-2 benchmarks at about 20 million instructions per second in the fast-forwarding mode. The benchmarks are compiled directly for a MIPS Linux system (no special SESC libraries).

+++ src/libcore/OSSim.h
@@ -185,7 +185,7 @@ public:
     if(idSimMarks.find(id)==idSimMarks.end()) {
       idSimMarks[id].total = 0;
       idSimMarks[id].begin = 0;
-      idSimMarks[id].end = (~0UL)-1;
+      idSimMarks[id].end = (uint)((~0UL)-1);
     }
 
     idSimMarks[id].total++;
